pyproject.toml
[project]
name = "assaia-qa-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "allure-pytest>=2.15.3",
    "playwright>=1.58.0",
    "pytest>=9.0.3",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.2",
]
